Move data sttructure defiitions to a single file
